Fix #359 - Improve performance for many islands of constraints.
The previous implementation iterated through the entire list of
constraints when solving each island. Now, the start index of each
island in m_sortedConstraints is stored when setup() is called.

Since the values for island ids are fairly sparse (the max island id is
the number of bodies minus 1, and the number of islands is usually much
smaller than that), I chose to use a hash table for storing the start
index / length of an island.